requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.18.0
pytube==10.8.5
typing-extensions==3.10.0.0
Werkzeug==1.0.1
youtube-dl==2021.4.26
